% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagmatrix.R
\name{lagmatrix}
\alias{lagmatrix}
\title{Create matrix with lead/lags of an input vector.}
\usage{
lagmatrix(x, lag)
}
\arguments{
\item{x}{Input variable.}

\item{lag}{Vector of leads and lags. Positive numbers are lags, negative are leads. O is the original x.}
}
\value{
\code{lmat} - Resulting matrix.
}
\description{
Create matrix with lead/lags of an input vector.
}
\examples{
  x <- rnorm(10)
  lagmatrix(x,c(0,1,-1))
}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
