% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnfor-package.R
\docType{package}
\name{nnfor-package}
\alias{nnfor-package}
\alias{_PACKAGE}
\title{nnfor: Time Series Forecasting with Neural Networks}
\description{
Automatic time series modelling with neural networks. Allows fully automatic, semi-manual or fully manual specification of networks. For details of the specification methodology see: (i) Crone and Kourentzes (2010) \doi{10.1016/j.neucom.2010.01.017}; and (ii) Kourentzes et al. (2014) \doi{10.1016/j.eswa.2013.12.011}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kourentzes.com/forecasting/2019/01/16/tutorial-for-the-nnfor-r-package/}
  \item Report bugs at \url{https://github.com/trnnick/nnfor/issues}
}

}
\author{
\strong{Maintainer}: Nikolaos Kourentzes \email{nikolaos@kourentzes.com}

}
\keyword{internal}
\keyword{package}
\keyword{ts}
