% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasdummy.R
\name{seasdummy}
\alias{seasdummy}
\title{Create seasonal dummy variables.}
\usage{
seasdummy(n, m = NULL, y = NULL, type = c("bin", "trg"), full = c(FALSE,
  TRUE))
}
\arguments{
\item{n}{Number of observations to create.}

\item{m}{Seasonal periodicity. If NULL it will take the information from the provided time series (y argument).}

\item{y}{This is an optional time series input that can be used to get seasonal periodicity (m) and the start point.}

\item{type}{Type of seasonal dummies to create. Can be "bin" for binary and "trg" for trigonometric.}

\item{full}{If full is TRUE then keeps the m-th dummy that is co-linear to the rest.}
}
\value{
\code{x} - Array with seasonal dummies.
}
\description{
Create binary or trigonometric seasonal dummies.
}
\examples{
  x <- seasdummy(24,12)

}
\author{
Nikolaos Kourentzes, \email{nikolaos@kourentzes.com}
}
